requirements.txt
python-dotenv~=1.0.1
langchain~=0.3.20
langchain-anthropic~=0.3.9
langchain-core~=0.3.44
langchain-openai~=0.3.8
langchain-community~=0.3.19
langchain-text-splitters~=0.3.6
pydantic~=2.10.6
langgraph~=0.3.7
requests~=2.32.3
openai~=1.66.2
llama-index-core~=0.12.23.post2
httpx~=0.28.1
uvicorn~=0.34.0
fastapi~=0.115.11
chromadb~=0.6.3
pytest~=8.3.5
langsmith~=0.3.13
mistralai~=1.5.1
langchain-experimental~=0.3.4
pandas~=2.2.3
openai-agents~=0.0.4